main
This is a new project stemming from trying to automate the action of sending out certificates. Most of the time, it requires a paid subscription. I have taken it upon myself to create a program that could do this for me. As of now I am working on allowing the program to be run from the Windows Terminal. For now, it is required that the source code be opened and modified accordingly for use.